[SCB-1956]rename project package to avoid random travis CI coverage error
