+++ demo/demo-local-registry/demo-local-registry-client/pom.xml
@@ -29,7 +29,7 @@
   </parent>
 
   <properties>
-    <demo.main>org.apache.servicecomb.demo.registry.Application</demo.main>
+    <demo.main>org.apache.servicecomb.demo.localRegistryClient.Application</demo.main>
   </properties>
   <dependencies>
     <dependency>
