+++ demo/demo-local-registry/demo-local-registry-client/src/main/java/org/apache/servicecomb/demo/localRegistryClient/Application.java
@@ -15,12 +15,9 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.demo.registry;
+package org.apache.servicecomb.demo.localRegistryClient;
 
-import java.util.concurrent.TimeUnit;
-
-import org.apache.servicecomb.demo.TestMgr;
-import org.apache.servicecomb.foundation.common.utils.BeanUtils;
+import org.apache.servicecomb.demo.CategorizedTestCaseRunner;
 import org.apache.servicecomb.springboot2.starter.EnableServiceComb;
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -28,17 +25,14 @@ import org.springframework.boot.builder.SpringApplicationBuilder;
 
 @SpringBootApplication
 @EnableServiceComb
-public class ServerApplication {
+public class Application {
   public static void main(final String[] args) throws Exception {
-    new SpringApplicationBuilder().sources(ServerApplication.class).web(WebApplicationType.SERVLET).build().run(args);
+    new SpringApplicationBuilder().sources(Application.class).web(WebApplicationType.SERVLET).build().run(args);
 
-    SelfServiceInvoker invoker = BeanUtils.getBean("SelfServiceInvoker");
-    invoker.latch.await(10, TimeUnit.SECONDS);
-    TestMgr.check(invoker.result, "hello");
+    runTest();
+  }
 
-    TestMgr.summary();
-    if (!TestMgr.errors().isEmpty()) {
-      System.exit(1);
-    }
+  public static void runTest() throws Exception {
+    CategorizedTestCaseRunner.runCategorizedTestCase("demo-local-registry-server");
   }
 }
