+++ demo/demo-local-registry/demo-local-registry-client/src/test/java/org/apache/servicecomb/demo/localRegistryClient/LocalRegistryIT.java
@@ -15,12 +15,13 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.demo.registry;
+package org.apache.servicecomb.demo.localRegistryClient;
 
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertThat;
 
 import org.apache.servicecomb.demo.TestMgr;
+import org.apache.servicecomb.demo.localRegistryClient.Application;
 import org.junit.Before;
 import org.junit.Test;
