+++ demo/demo-local-registry/demo-local-registry-server/pom.xml
@@ -29,7 +29,7 @@
   </parent>
 
   <properties>
-    <demo.main>org.apache.servicecomb.demo.registry.ServerApplication</demo.main>
+    <demo.main>org.apache.servicecomb.demo.localRegistryServer.ServerApplication</demo.main>
   </properties>
   <dependencies>
     <dependency>
